updated methods

+++ AMT-Project/src/main/java/ch/heigvd/amt/project/model/Field.java
@@ -6,6 +6,6 @@ import lombok.*;
 @EqualsAndHashCode
 @Getter
 public class Field {
-    private long idField;
+    private int idField;
     private int size;
 }
